Skip basic scenario test
This test is very frequently failing now. The test case creates a stack
with an instance but the stack can't be deleted within timeout because
of the known libvirt bug in Ubuntu Jammy. Because the release deadline
is already over, we'll disable this test so that we can merge some
changes now. We'll revisit this (and the other test cases we disabled)
later.

Related-Bug: #1998274
Story: 2010631
Task: 47589
Change-Id: I87c4b1e7a911fd78e327393b1af200667e89c999