Add non-voting CentOS 9 Stream job
This introduces non-voting CentOS 9 Stream functional jobs, to restore
some tests which were disabled because of the known libvirt bug. Though
the bug also affects CentOS 9 Stream, we observe much low failure rate
in CentOS 9 Stream so we can restore these test cases in CentOS 9
Stream. The job is kept non-voting until it is proven to be stable
after a while.

Change-Id: I37211aa941be1892ad0ddf0694a758972a0aebba