Deprecate heat-all
It was added to be used in TripleO, which was already retired. Because
we do not expect any other use case in real deployments, deprecate it
so that we can remove it. The current implementation heavily relies on
the fact that we have wsgi server implementation, which is no longer
available once eventlet is removed.

Change-Id: I8c177b63350553a30c52d51b878a5a5473ccefb8