Fix stack update when reauthentication_auth_method=trusts
When a stack is created with a certain user,
"reauthentication_auth_method=trusts" and an update is
triggered, the update fails because after the trust
regeneration the user_creds_id values are deleted for
the stack and the nested resources. Only the parent
stack gets a new set of creds but those are not
propagated to the nested resources.

This patch grabs the user_creds_id created for the
parent stack upon trust regeneration and propagates
that to the nested stacks.

story: 2011342
task: 51653

Change-Id: I9a3263c6ab3d515fa468cfb7ce3cd6cfbf39aa57
Signed-off-by: Alejandro Santoyo <alejandro.santoyo@canonical.com>