Remove redundant monkey patch
eventlet.monkey_patch patches default modules unless only specific
modules are requested to be patched. (In case specific modules are
requested to be non-patched, then it patches the other modules in
the default set).

Because the time module is always patched in the base commands, we
don't have to explicitly patch it in run time. Note that we still need
monkey patch in tests so patching is migrated to the test code.

Change-Id: I3f6a57bdadded3473840aee975329ebc505d9af6