Deprecate console scripts for api services
Current standalone api services heavily depend on WSGI server
implementation of eventlet and we aim to completely get rid of it as
part of the effort to get rid of eventlet.

It is now officially recommended to use the wsgi scripts or
applications and run these by a 3rd party mechanism such as uwsgi or
httpd + mod_wsgi.

Change-Id: I91457df91bb2e6f6710af6f4fc19b11527d8e4c3