Fix check_interface_detach()
The os-interface API is dependent on device_id set in the port in
neutron port API call, and does not return a port if neutron completed
unbinding the port but nova hasn't yet completed its detachment
process. If we attempt to attach the port to a different instance in
this situation, nova rejects it because it does not allow having two
VIF records with the same mac. To avoid this timing problem, we use
the addresses field in nova server API, which returns the information
according to the VIF records kept in nova database.

Note that the addresses field does not contain the information to
identify the exact port being detached. So the detachment check may
wait forever in case
 - A server has two ports attached
 - The two ports are allocated from different networks
 - The ports get the same mac address allocated
due to the fact that neutron allows allocating the same mac address
for different networks, which may rarely happen. In this case the port
has to be deleted manually.

Change-Id: Ia69603c8ffe85a9b29bb699a3dea560516f82d40
Signed-off-by: rabi <ramishra@redhat.com>